value-relation.h: Mark dom_oracle::next_relation as override
When GCC is compiled with clang, it emits a warning that
dom_oracle::next_relation is not marked as override even though it
does override a virtual function of its ancestor.  This patch marks it
as such to silence the warning and for the sake of consistency.

There are other member functions in the class which are marked as
final override but this particular function is in the protected
section so I decided to just mark it as override.

gcc/ChangeLog:

2025-06-24  Martin Jambor  <mjambor@suse.cz>

	* value-relation.h (class dom_oracle): Mark member function
	next_relation as override.