Add testcase for afdo offlining and fix two bugs
This patch adds a testcase that offlining works and profile info is not lost.
While doing it I noticed a pasto that made the dump to be "afdo" and not
"afdo_offline" and also that not all functions are processed as the range
for does not expect new values to be put to the vector.  Fixed thus.

gcc/ChangeLog:

	* auto-profile.cc (function_instance::merge): Add TODO.
	(autofdo_source_profile::offline_external_functions):
	Do not use range for on the worklist.
	* timevar.def (TV_IPA_AUTOFDO_OFFLINE): New timevar.

gcc/testsuite/ChangeLog:

	* gcc.dg/tree-prof/afdo-crossmodule-1.c: New test.
	* gcc.dg/tree-prof/afdo-crossmodule-1b.c: New test.