Fixup vector epilog analysis skipping when not using partial vectors
The following avoids re-analyzing the loop as epilogue when not
using partial vectors and the mode is the same as the autodetected
vector mode and that has a too high VF for a non-predicated loop.
This situation occurs almost always on x86 and saves us one
re-analysis unless --param vect-partial-vector-usage is non-default.

	* tree-vectorizer.h (vect_chooses_same_modes_p): New
	overload.
	* tree-vect-stmts.cc (vect_chooses_same_modes_p): Likewise.
	* tree-vect-loop.cc (vect_analyze_loop): Prune epilogue
	analysis further when not using partial vectors.