[lra] inactivate disabled fp2sp elimination [PR120424]
Even after we disable the fp2sp elimination when it is the active
elimination for the fp, spilling might use it before
update_reg_eliminate runs and inactivates it for good.  If it is used,
update_reg_eliminate will fail the check that fp2sp was not used.

Since we keep track of uses of this specific elimination, and
lra_update_fp2sp_elimination checks it before disabling it, we know it
hasn't been used, so we can inactivate it without any ill effects.

This fixes the pr118591-1.c avr-none regression exposed by the
PR120424 fix.


for  gcc/ChangeLog

	PR rtl-optimization/120424
	* lra-eliminations.cc (lra_update_fp2sp_elimination):
	Inactivate the unused fp2sp elimination right away.