Impove diagnostics of mismatched discriminators in auto-profile
We are missing discriminator info in auto-profiles, for example in exchange2.
I am not sure why, since I see the info still present in dwarf2out, so it may
be bug at create_gcov side.

This patch makes the workaround to ouptput better diagnostics (to actually show
the soruce location).  This needs promotion of location info through the inline
stack API, so I turned it from pair to actual structure.  Overall I think pairs
are overused in this source and makes it harder to read.

Bootstrapped/regtested x86_64-linux, comitted.

gcc/ChangeLog:

	* auto-profile.cc (struct decl_lineno): Turn to structure; add
	location.
	(dump_inline_stack): Update.
	(get_inline_stack): Update.
	(get_relative_location_for_locus): Fixup formating.
	(function_instance::get_function_instance_by_decl): Add
	LOCATION parameter; improve dumping.
	(autofdo_source_profile::get_callsite_total_count): Improve dumping;
	update.
	(walk_block): Update.
	(autofdo_source_profile::offline_unrealized_inlines): Update.
	(autofdo_source_profile::get_count_info): Update.