testsuite: Fix up pr119318.c test for big-endian [PR120082]
The test is not endianess clean, x[0] is supposed to be ((__int128)0x19)<<32
on little endian - 0x19 is in the second vector elt - but ((__int128)0x19)<<64
on big endian.  I've added also verification of int and __int128 sizes just
in case we have say 16-bit or 64-bit int target with __int128 type, or
pdp endian gets __int128 support.

2025-07-01  Jakub Jelinek  <jakub@redhat.com>

	PR ipa/119318
	PR testsuite/120082
	* gcc.dg/ipa/pr119318.c (main): Expect different result on big endian
	from little endian, on unexpected endianness or int/int128 sizes don't
	test anything.  Formatting fixes.