tree-optimization/118669 - fixup wrongly aligned loads/stores
The vectorizer tracks alignment of datarefs with dr_aligned
and dr_unaligned_supported but that's aligned with respect to
the target alignment which can be less aligned than the mode
used for the access.  The following fixes this discrepancy
for vectorizing loads and stores.  The issue is visible for
aarch64 SVE and risc-v where VLA vector modes have larger than
element alignment but the target handles element alignment
just fine.

	PR tree-optimization/118669
	* tree-vect-stmts.cc (vectorizable_load): Emit loads
	with proper (element) alignment.
	(vectorizable_store): Likewise.