Fortran: Remove corank conformability checks [PR120843]
Remove the checks on coranks conformability in expressions,
because there is nothing in the standard about it.  When a coarray
has no coindexes it it treated like a non-coarray, when it has
a full-corank coindex its result is a regular array.  So nothing
to check for corank conformability.

	PR fortran/120843

gcc/fortran/ChangeLog:

	* resolve.cc (resolve_operator): Remove conformability check,
	because it is not in the standard.

gcc/testsuite/ChangeLog:

	* gfortran.dg/coarray/coindexed_6.f90: Enhance test to have
	coarray components covered.

(cherry picked from commit 15413e05eb9cde976b8890cd9b597d0a41a8eb27)