c++: -Wno-abbreviated-auto-in-template-arg [PR120917]
In r14-1659 I added a missing error for a Concepts TS feature that we were
failing to diagnose, but this PR requests a way to disable that error for
code written thinking it was valid.  Which seems reasonable, since it
doesn't require any work beyond that and is a plausible extension by itself.

While looking at this, I also noticed we were still not giving the
diagnostic in a few cases, and fixing that affected a few of our old
concepts testcases.

	PR c++/120917

gcc/ChangeLog:

	* doc/invoke.texi: Add -Wno-abbreviated-auto-in-template-arg.

gcc/c-family/ChangeLog:

	* c.opt: Add -Wno-abbreviated-auto-in-template-arg.
	* c.opt.urls: Regenerate.

gcc/cp/ChangeLog:

	* parser.cc (cp_parser_simple_type_specifier): Attach
	auto in targ in parameter to -Wabbreviated-auto-in-template-arg.
	(cp_parser_placeholder_type_specifier): Diagnose constrained auto in
	template arg.

gcc/testsuite/ChangeLog:

	* g++.dg/concepts/auto7a.C: Add diagnostic.
	* g++.dg/concepts/auto7b.C: New test.
	* g++.dg/concepts/auto7c.C: New test.
	* g++.dg/cpp1y/pr85076.C: Expect 'auto' error.
	* g++.dg/concepts/pr67249.C: Likewise.
	* g++.dg/cpp1y/lambda-generic-variadic.C: Likewise.
	* g++.dg/cpp2a/concepts-pr67210.C: Likewise.
	* g++.dg/concepts/pr67249a.C: New test.
	* g++.dg/cpp1y/lambda-generic-variadic-a.C: New test.
	* g++.dg/cpp2a/concepts-pr67210a.C: New test.