Do not query further vector epilogues after a masked epilogue
When doing --param vect-partial-vector-usage=1 we'd continue querying
the target whether it wants more vector epilogues, but when it comes
back with a suggestion we then might iterate endlessly.  Do not
even ask the target when we decided for the last epilogue to be
one with partial vectors.

	PR tree-optimization/120927
	* tree-vect-loop.cc (vect_analyze_loop): Stop querying
	further epilogues after one with partial vectors.

(cherry picked from commit 57d30924e2b3887a1aedc6102daa67bc3272ace7)